# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delicious Dinner API",
        "description": "API realizada en Laravel",
        "contact": {
            "email": "ingisraeltrujillo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/metodo_pago": {
            "get": {
                "tags": [
                    "metodo_pago"
                ],
                "summary": "Muestra todos los metodos de pago",
                "description": "Muestra todos los metodos de pago.",
                "operationId": "40ee9c2a49b6d3ef89f07425c954d5c9",
                "responses": {
                    "200": {
                        "description": "Muestra todos los metodos de pago"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "metodo_pago"
                ],
                "summary": "Crear un nuevo metodo de pago",
                "operationId": "483cf3691dc2b0ab453265585d0f943f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "EFECTIVO",
                                    "codename": "cash"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    }
                }
            }
        },
        "/api/metodo_pago/{id}": {
            "get": {
                "tags": [
                    "metodo_pago"
                ],
                "summary": "Mostrar la información de un metodo de pago",
                "operationId": "cd231a78e907474e479ef247989b7798",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "metodo_pago"
                ],
                "summary": "Actualizar un metodo de pago",
                "operationId": "2ea34f9389aadc364ee224d050fe2e90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "EFECTIVO",
                                    "codename": "cash"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "metodo_pago"
                ],
                "summary": "Eliminar un metodo de pago",
                "operationId": "5f32d74dcc63a75533bd819c13bd395a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}